requirements.txt
requests>=2.31
feedparser>=6.0
pyyaml>=6.0
flask>=3.0
python-dateutil>=2.8
supabase>=2.0
python-dotenv>=1.0
gunicorn>=21.0
